moving of bucket/object resolving impl
